Statement caching: cleanup
