+++ db/comdb2.h
@@ -146,12 +146,6 @@ enum SYNC_FLAGS {
     REP_SYNC_N = 4 /* wait for N machines */
 };
 
-enum STMT_CACHE_FLAGS {
-    STMT_CACHE_NONE = 0,
-    STMT_CACHE_PARAM = 1,
-    STMT_CACHE_ALL = 2 /* cache all the queries */
-};
-
 enum OPCODES {
     OP_DBINFO = 0 /*rmtdb info req*/
     ,
@@ -1013,6 +1007,7 @@ struct thread_info {
     void *ct_add_table;
     void *ct_del_table;
     void *ct_add_index;
+    void *stmt_cache;
 };
 
 /* key for fstblk records - a 12 byte sequence number generated by the
